// src/app/routes/routes.enum.ts
export enum ROUTES {
  ROOT = '/',
  PROJECT = '/project',
  OBJECT = '/object',
  RD = '/rd',
  MTO = '/mto',
  CMP = '/cmp',
  GRAPH = '/graph',
  MIM = '/mim',
  WORKERS = '/workers',
  CAPITAL = '/capital',
  BUDGET = '/budget',
  FINANCING = '/financing',
  PANORAMS = '/panorams',
  CAMERAS = '/cameras',
  ASSIGNMENTS = '/assignments',
  CONTRACTORS = '/contractors',
}